% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{chisq_test}
\alias{chisq_test}
\title{A function to compute the chi-square test statistic when the input is two numeric vectors}
\usage{
chisq_test(x, y)
}
\arguments{
\item{x}{the first input vector}

\item{y}{the second input vector}
}
\value{
the chi-square test statistic
}
\description{
A function to compute the chi-square test statistic when the input is two numeric vectors
}
\examples{
\dontrun{
rnR <-chisq_test(c(3,8,4,9),c(6,9,3,5))
}
}
