% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{Jacknife.estimate}
\alias{Jacknife.estimate}
\title{A function to estimate the bias and the standard error of the correlation statistic by jacknife method}
\usage{
Jacknife.estimate(x, y)
}
\arguments{
\item{x}{the first input variable}

\item{y}{the second input variable}
}
\value{
the estimate of the bias and the standard error of the correlation statistic
}
\description{
A function to estimate the bias and the standard error of the correlation statistic by jacknife method
}
\examples{
\dontrun{
rnR <-Jacknife.estimate(c(0.1,0.5,1,1.2,0.2,0.4,0.8,0.6,0.8,1.3),c(4,2.7,5.4,3.6,6.3,3.9,4.7,3.9,2.8,5.6))
}
}
