% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.R
\name{MCestimate}
\alias{MCestimate}
\title{A Monte Carlo estimate of the Beta(3,3) using R}
\usage{
MCestimate(n, x)
}
\arguments{
\item{n}{the replicates}

\item{x}{the number of between-sample random numbers}
}
\value{
the estimated values returned by the my_MCestimate function and the value computated by the cdf
}
\description{
A Monte Carlo estimate of the Beta(3,3) using R
}
\examples{
\dontrun{
rnR <- MCestimate(1000,c(1,2,3,4,5,6,7,8,9)/10)
}
}
